BUG: Fix float-precision division in BresenhamLine::BuildLine
Add explicit static_cast<double>() on both operands of the division
computing euclideanLineLen. Because LType = Vector<float>, the
expression `IdentifierType / float` previously evaluated in float
precision; the double result was only a widened float quotient.

+++ Modules/Core/Common/include/itkBresenhamLine.hxx
@@ -41,7 +41,8 @@ BresenhamLine<VDimension>::BuildLine(LType Direction, IdentifierType length) -> 
   // compute actual line length because the shorter distance
   // the larger deviation due to rounding to integers
   const IdentifierType mainDirectionLen = length - 1;
-  const double         euclideanLineLen = mainDirectionLen / itk::Math::Absolute(Direction[maxDistanceDimension]);
+  const double         euclideanLineLen =
+    static_cast<double>(mainDirectionLen) / static_cast<double>(itk::Math::Absolute(Direction[maxDistanceDimension]));
 
   // we are going to start at 0
   constexpr IndexType StartIndex{ 0 };
